chore(shield): move `ast` outside internal folder
